Opportunistic Cached Authentication v2
This is what rfc2617 says.

  A client SHOULD assume that all paths at or deeper than the depth of
  the last symbolic element in the path field of the Request-URI also
  are within the protection space specified by the Basic realm value of
  the current challenge. A client MAY preemptively send the
  corresponding Authorization header with requests for resources in
  that space without receipt of another challenge from the server.

  Similarly, when a client sends a request to a proxy, it may reuse a
  userid and password in the Proxy-Authorization header field without
  receiving another challenge from the proxy server.

Use Gisle's idea of handlers in UserAgent, but non-host-specific
because we want to be able to lookup up the proxy or the host at
use time.

Also, support digest 'domain'.

NOTE: the only changes to tests have been to make it work again by not
reusing $req objects that have had headers inserted already.  There are
no tests that this behaviour actually works.  It's a bit tricky to test
without forcing the UA not to re-try with the authentication details
added.

Bron.